requirements.txt
discord.py
redispy
jishaku
better-profanity
youtube_dl
PyNaCl
numpy
lyricsgenius
spotipy
Pillow
psutil